feat(migration): Implement category, coupon, and course migration services
- Added CategoryMigrationService to create categories from JSON data, ensuring unique slugs.
- Introduced CouponMigrationService to seed coupon data with various discount types and conditions.
- Developed CourseMigrationService to read course data from JSON, create courses, sections, and associated content.
- Created QuizMockDataService for generating quiz questions and answers tailored to courses.
- Added utility classes for generating course and reading content dynamically based on category.
- Implemented VideoData utility for selecting appropriate video titles and paths based on course categories.
